Feature/setup contracts and tests (#5)
* Setup configs

* Initial code commits

* Added identity contract and updated test cases

* Remove unused file

* Removed unused file

* Setting up contracts

* Added the math util

* Renamed math util

* Renamed math util

* adding utils

+++ contracts/Token.sol
@@ -9,7 +9,7 @@ pragma solidity ^0.8.17;
 contract Token {
     string public name = 'Sargo Token';
     string public symbol = 'SGT';
-    uint public totalSupply = 1000000;
+    uint256 public totalSupply = 1000000;  // * 10 ** 18
     address public owner;
     mapping(address => uint) balances;
